Split out annotation parsers into a separate file
